mqtt-ingestor/requirements.txt
paho-mqtt==1.6.1
confluent-kafka==2.6.1
six==1.16.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0